librenderengine: Fix -Wformat errors
frameworks/native/libs/renderengine/skia/AutoBackendTexture.cpp:80:51: error: format specifies type 'int' but the argument has type 'GrBackendApi' [-Werror,-Wformat]
   80 |         LOG_ALWAYS_FATAL("Unexpected backend %d", backend);
      |                                              ~~   ^~~~~~~
      |                                                   static_cast<int>(
system/logging/liblog/include/log/log_main.h:142:48: note: expanded from macro 'LOG_ALWAYS_FATAL'
  142 |   (((void)android_printAssert(NULL, LOG_TAG, ##__VA_ARGS__)))
      |                                                ^~~~~~~~~~~
system/logging/liblog/include/log/log_main.h:125:43: note: expanded from macro 'android_printAssert'
  124 |                        __android_second(0, ##__VA_ARGS__, NULL) \
      |                                              ~~~~~~~~~~~
  125 |                            __android_rest(__VA_ARGS__))
      |                                           ^~~~~~~~~~~
system/logging/liblog/include/log/log_main.h:120:40: note: expanded from macro '__android_rest'
  120 | #define __android_rest(first, ...) , ##__VA_ARGS__
      |                                        ^~~~~~~~~~~
frameworks/native/libs/renderengine/skia/AutoBackendTexture.cpp:148:50: error: format specifies type 'int' but the argument has type 'ui::Dataspace' [-Werror,-Wformat]
  144 |             LOG_ALWAYS_FATAL("%s isTextureValid:%d dataspace:%d"
      |                                                              ~~
  145 |                              "\n\tGrBackendTexture: (%i x %i) hasMipmaps: %i isProtected: %i "
  146 |                              "texType: %i\n\t\tVkImageInfo: success: %i fFormat: %i "
  147 |                              "fSampleCount: %u fLevelCount: %u colorType %i",
  148 |                              msg, tex.isValid(), dataspace, tex.width(), tex.height(),
      |                                                  ^~~~~~~~~
      |                                                  static_cast<int>(
system/logging/liblog/include/log/log_main.h:142:48: note: expanded from macro 'LOG_ALWAYS_FATAL'
  142 |   (((void)android_printAssert(NULL, LOG_TAG, ##__VA_ARGS__)))
      |                                                ^~~~~~~~~~~
system/logging/liblog/include/log/log_main.h:125:43: note: expanded from macro 'android_printAssert'
  124 |                        __android_second(0, ##__VA_ARGS__, NULL) \
      |                                              ~~~~~~~~~~~
  125 |                            __android_rest(__VA_ARGS__))
      |                                           ^~~~~~~~~~~
system/logging/liblog/include/log/log_main.h:120:40: note: expanded from macro '__android_rest'
  120 | #define __android_rest(first, ...) , ##__VA_ARGS__

Bug: 315250603
Test: m librenderengine

+++ libs/renderengine/skia/AutoBackendTexture.cpp
@@ -77,7 +77,7 @@ AutoBackendTexture::AutoBackendTexture(GrDirectContext* context, AHardwareBuffer
                                                                  backendFormat,
                                                                  isOutputBuffer);
     } else {
-        LOG_ALWAYS_FATAL("Unexpected backend %d", backend);
+        LOG_ALWAYS_FATAL("Unexpected backend %d", static_cast<int>(backend));
     }
 
     mColorType = GrAHardwareBufferUtils::GetSkColorTypeFromBufferFormat(desc.format);
@@ -145,8 +145,8 @@ void logFatalTexture(const char* msg, const GrBackendTexture& tex, ui::Dataspace
                              "\n\tGrBackendTexture: (%i x %i) hasMipmaps: %i isProtected: %i "
                              "texType: %i\n\t\tVkImageInfo: success: %i fFormat: %i "
                              "fSampleCount: %u fLevelCount: %u colorType %i",
-                             msg, tex.isValid(), dataspace, tex.width(), tex.height(),
-                             tex.hasMipmaps(), tex.isProtected(),
+                             msg, tex.isValid(), static_cast<int>(dataspace), tex.width(),
+                             tex.height(), tex.hasMipmaps(), tex.isProtected(),
                              static_cast<int>(tex.textureType()), retrievedImageInfo,
                              imageInfo.fFormat, imageInfo.fSampleCount, imageInfo.fLevelCount,
                              colorType);
